Major program refactor (#159)
* Entry <--> EntryChunk <--> NSF links

* finish connections and stuff

* new NSF entyr lookup

* test new live EID lookup

* OldAnim viewer works

* anchor mode

* lighting proto

* cut lighting for now

* CrashEdit: use latest C# lang version

* texturing attempt (broken)

* update packages to match ce2 (also i need the updates anyway)

* enable debug stuff

* fix crazy cpu usage

* attempt two

* it works holy fk

* textures work properly

* interp works

* fixes

* cull fixes

* fix memory leak with shaders

* support transparent texels

* render sync fixes & improve interp considerably

* perf hack for 1-frame anims

* blend mode support proto

* display collision

* cvtx works

* move stuff

* render pass for trans texels

* texture toggle!

* rename functions

* small perf fix

* Update OldAnimationEntryViewer.cs

* automatically detect shaded vs colored anims

* C1 scenery

* move user params to VAO and merge crash 1 renderers

* optimize rendering more

* revamp interp system to be less janky

* fix bug

* render all worlds at once (slight perf hit probably)

* proto scenery works + new option

* generate a resource texture atlas at startup

* temp crash 1 zone viewer

* random changes

* Update SoundBox.cs

* refactor some things

* attempt to only use a single GL context for all GLViewers

* make a method for loading GL controls in static space

* rename DummyTopGLViewer -> GLViewerLoader

* fix project

* test multi-pass VAO and unpacked tex struct

* testing speed

* use single VAO again for now

* finish optimizing OldSceneryEntryViewer

* optimize ProtoSceneryEntryViewer

* optimize OldAnimationEntryViewer

* rewrite `OldZoneEntryViewer`

* make "show all level..." options not prone to leaks

* remove more unused code

* Add optimized sprite renderer + improve all other renderers

* also old sprite renderer

* proper sky rendering (add toggle later)

* use indices for the static models

* octree renderer reimplementation

* cleanup and further optimizations

* use rgba struct in more places

* render grid after everything + make octree renderer have changeable alpha

* add normal renderer

* fix multi-zone renderer when there's only 1 zone in list

* get rid of annoying uniforms

* renamed some fields and sprite shaders

* fix sky rendering

* Update MathExt.cs

* fix `SetBlendMode` accidentally overriding z-write in sky renderer

* make cameras show their facing angle

* remove a couple of useless calls

* more + fix boxes always having max alpha

* fix crash on close

* make proto zone/anim viewers work

* crash 2 scenery renderer

* crash 2 zone renderer

* show cameras in crash 2 zone viewer

* fix position of boxes in c2 zone viewer + simpler camera code

* Update GLViewer.cs

* add check before rendering camera angles for crash 2 zone viewer

* Increase depth buffer resolution from 16bits to 24 bits

* crash 2 animation viewer

* reuse more VAOs

* fixes

* update packages

* Fix memory leak in `RenderInfo`

* change zoom params a bit

* formatting

* linting

* render uncompressed vertices in crash 2 anim viewer

* rename some fields

* more cleanup

* fixes...

* text rendering test success

* Font settings

* Also grab user fonts

* make `GLLoadStatic` work properly

* Fix dumb zbuf access bugs

* basic 3D text rendering

* Fix font selector

* Add 3D text scaling and text shadow

* Redesign settings panel for new settings

* Localize a string

* Fix font setting checkboxes not being set at startup

* More debug render time info

* fix orientation in box verts

* Update parameters for 3D controls

* also show entity names for entities with paths

* also render zone name!

* add text alignment to text renderer

* text fixes

* render gem ID for, well, gems.

* also apply alpha to text shadow

* fix animation centering

* Update GLViewer.cs

* fix vao zbuffer settings + fix depth for zone viewers

* add a bunch of extra displays to crash 2 zone renderer

* add separate thick line renderer

* fix some bad sceneries breaking things

* tweak zoom speed

* rewrite some rgba code

* add fancy text displays to all crash 1 versions

* Also allow .otf font files.

* more font diagnostics

* fix annoying sprite flicker

* fix maps not loading properly

* fix generic tris not rendering oops

* make thick lines thicker

* make `IPosition` use floats for better storage + speed and remove it from various structs

* Delete Crash 3 duplicare files

* fix zone renderer for crash 3

* better color for yellow gem

* minor changes and support weird box object

* Fix crash 3 sky renderer

* Crash 3 animation renderer (crappy as always)

* fix crash 3 sky offset

* little temp hack for skies with depth

* Text alignment fixes + render help text

* Use a common interface for setting up keyboard controls and help text

* settings UI for help display

* Localize.

* Localize more text.

* minor change to zone renderer

* rename shaders

* time trial mode switch

* code cleanup

* fix crash 2 scenery

* more cleanup and code duplication removing

* unmaintained file

* Update Program.cs

* make transparent polys render slightly better

* fix crash 1 normal rendering

* cleanup some `OldFrame` things

* run cleanup on `Crash` so i can stop fighting vs

* update version!

* fix up appveyor script and drop old DLL's

* fix typo in appveyor script

* fix appveyor script natives dir path

* fix crash 3 NSDs

* create console window first

* revamp collision viewer

* useless usings

* minor refactor

* add 32bit and 64bit configurations just for testing

* fix outline octree renderer bug

* minor appveyor fixes

---------

+++ Crash/Utility/IPosition.cs
@@ -2,8 +2,8 @@ namespace Crash
 {
     public interface IPosition
     {
-        double X { get; }
-        double Y { get; }
-        double Z { get; }
+        float X { get; }
+        float Y { get; }
+        float Z { get; }
     }
 }
